Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -7,6 +7,6 @@ char* welcome() {
 int main() {
     int a,b,c;
     scanf("%d ",&a);scanf("%d ",&b);scanf("%d ",&c);
-    int avg = a+b+c/3;
+    float avg = a+b+c/3;
     printf("%d",avg);
 }
